# api/openapi.yaml
components:
  schemas:
...
    ErrorBody:
      additionalProperties: true
      properties:
        code:
          type: string
        details: {}
        message:
          type: string
      required:
        - code
        - message
      type: object
    ErrorEnvelope:
      additionalProperties: true
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
      required:
        - error
      type: object
    ExecutionDetail:
      additionalProperties: true
      properties:
        chain_depth:
          format: int64
          type: integer
        children:
          items:
            $ref: "#/components/schemas/ExecutionRef"
          type: array
        created_at:
          format: date-time
          type: string
        created_by:
          type: string
        duration_ms:
          format: int64
          type:
            - integer
            - "null"
        ended_at:
          format: date-time
          type:
            - string
            - "null"
        error:
          type: string
        flow_id:
          description: Flow ID. Null for file runs.
          type:
            - string
            - "null"
        flow_revision_id:
          type:
            - string
            - "null"
        git_sha:
          type: string
        id:
          type: string
        inputs:
          additionalProperties: {}
          type: object
        labels:
          additionalProperties:
            type: string
          type: object
        namespace:
          type: string
        outputs:
          additionalProperties: {}
          type:
            - object
            - "null"
        parent_execution_id:
          type:
            - string
            - "null"
        reason:
          type: string
        restart_of_id:
          type:
            - string
            - "null"
        secret_keys_used:
          items:
            type: string
          type: array
        snapshot_id:
          type: string
        snapshot_version:
          format: int64
          type:
            - integer
            - "null"
        started_at:
          format: date-time
          type:
            - string
            - "null"
        state:
          enum:
            - QUEUED
            - RUNNING
            - CANCELLING
            - SUCCESS
            - FAILED
            - TIMED_OUT
            - CANCELLED
            - SKIPPED
          type: string
        task_runs:
          items:
            $ref: "#/components/schemas/TaskRun"
          type: array
        trigger_payload:
          additionalProperties: {}
          type: object
        trigger_type:
          type: string
      required:
        - inputs
        - trigger_payload
        - snapshot_id
        - chain_depth
        - secret_keys_used
        - task_runs
        - children
        - id
        - namespace
        - state
        - trigger_type
        - labels
        - created_at
        - error
        - reason
        - created_by
      type: object
    ExecutionList:
      additionalProperties: true
      properties:
        items:
          items:
            $ref: "#/components/schemas/ExecutionSummary"
          type: array
        next_cursor:
          type: string
      required:
        - items
      type: object
    ExecutionRef:
      additionalProperties: true
      properties:
        created_at:
          format: date-time
          type: string
        id:
          type: string
        state:
          type: string
      required:
        - id
        - state
        - created_at
      type: object
    ExecutionSummary:
      additionalProperties: true
      properties:
        created_at:
          format: date-time
          type: string
        created_by:
          type: string
        duration_ms:
          format: int64
          type:
            - integer
            - "null"
        ended_at:
          format: date-time
          type:
            - string
            - "null"
        error:
          type: string
        flow_id:
          description: Flow ID. Null for file runs.
          type:
            - string
            - "null"
        id:
          type: string
        labels:
          additionalProperties:
            type: string
          type: object
        namespace:
          type: string
        reason:
          type: string
        started_at:
          format: date-time
          type:
            - string
            - "null"
        state:
          enum:
            - QUEUED
            - RUNNING
            - CANCELLING
            - SUCCESS
            - FAILED
            - TIMED_OUT
            - CANCELLED
            - SKIPPED
          type: string
        trigger_type:
          type: string
      required:
        - id
        - namespace
        - state
        - trigger_type
        - labels
        - created_at
        - error
        - reason
        - created_by
      type: object
...
    LogEntry:
      additionalProperties: true
      properties:
        attempt:
          format: int64
          type: integer
        "n":
          description: 1-based line number in the task run.
          format: int64
          type: integer
        stream:
          enum:
            - stdout
            - stderr
            - system
          type: string
        task_key:
          type: string
        task_run_id:
          type: string
        text:
          type: string
        ts:
          format: date-time
          type: string
      required:
        - task_run_id
        - task_key
        - attempt
        - "n"
        - ts
        - stream
        - text
      type: object
    LogPage:
      additionalProperties: true
      properties:
        done:
          description: True when the execution ended and no more lines follow.
          type: boolean
        lines:
          items:
            $ref: "#/components/schemas/LogEntry"
          type: array
        next_cursor:
          type: string
      required:
        - lines
        - done
      type: object
...
    RunFileRequest:
      additionalProperties: true
      properties:
        args:
          items:
            type: string
          maxItems: 100
          type: array
        path:
          maxLength: 512
          type: string
      required:
        - path
      type: object
    RunnerComplete:
      additionalProperties: true
      properties:
        error:
          type: string
        exit_code:
          format: int64
          type: integer
        reason:
          type: string
      required:
        - exit_code
        - error
      type: object
    RunnerEvent:
      additionalProperties: true
      properties:
        key:
          type: string
        name:
          type: string
        tags:
          additionalProperties:
            type: string
          type: object
        ts:
          format: date-time
          type: string
        type:
          enum:
            - output
            - metric
          type: string
        unit:
          type: string
        value: {}
      required:
        - type
      type: object
    RunnerEventBatch:
      additionalProperties: true
      properties:
        events:
          items:
            $ref: "#/components/schemas/RunnerEvent"
          maxItems: 1000
          type: array
        seq:
          format: int64
          minimum: 1
          type: integer
      required:
        - seq
        - events
      type: object
    RunnerHeartbeatResponse:
      additionalProperties: true
      properties:
        cancel:
          type: boolean
      required:
        - cancel
      type: object
    RunnerLimits:
      additionalProperties: true
      properties:
        max_artifact_bytes:
          format: int64
          type: integer
        max_bundle_bytes:
          format: int64
          type: integer
      required:
        - max_artifact_bytes
        - max_bundle_bytes
      type: object
    RunnerLogBatch:
      additionalProperties: true
      properties:
        lines:
          items:
            $ref: "#/components/schemas/RunnerLogLine"
          type: array
        seq:
          format: int64
          minimum: 1
          type: integer
      required:
        - seq
        - lines
      type: object
    RunnerLogLine:
      additionalProperties: true
      properties:
        stream:
          enum:
            - stdout
            - stderr
            - system
          type: string
        text:
          type: string
        ts:
          format: date-time
          type: string
      required:
        - ts
        - stream
        - text
      type: object
    RunnerSpec:
      additionalProperties: true
      properties:
        attempt:
          format: int64
          type: integer
        bundle_hash:
          type: string
        command:
          items:
            type: string
          type: array
        env:
          additionalProperties:
            type: string
          type: object
        execution_id:
          type: string
        flow_id:
          type: string
        limits:
          $ref: "#/components/schemas/RunnerLimits"
        mask_values:
          items:
            type: string
          type: array
        namespace:
          type: string
        runtime:
          type: string
        task_id:
          type: string
        task_run_id:
          type: string
        timeout_seconds:
          format: int64
          type: integer
        workdir:
          type: string
      required:
        - task_run_id
        - execution_id
        - namespace
        - flow_id
        - task_id
        - attempt
        - command
        - workdir
        - env
        - timeout_seconds
        - mask_values
        - bundle_hash
        - limits
      type: object
...
    TaskRun:
      additionalProperties: true
      properties:
        attempt:
          format: int64
          type: integer
        child_execution_id:
          type:
            - string
            - "null"
        duration_ms:
          format: int64
          type:
            - integer
            - "null"
        ended_at:
          format: date-time
          type:
            - string
            - "null"
        error:
          type: string
        executor_type:
          type: string
        exit_code:
          format: int64
          type:
            - integer
            - "null"
        id:
          type: string
        outputs:
          additionalProperties: {}
          type:
            - object
            - "null"
        pool:
          type: string
        queued_at:
          format: date-time
          type:
            - string
            - "null"
        reason:
          type: string
        reused_from_id:
          type:
            - string
            - "null"
        started_at:
          format: date-time
          type:
            - string
            - "null"
        state:
          enum:
            - PENDING
            - QUEUED
            - RUNNING
            - SUCCESS
            - FAILED
            - TIMED_OUT
            - CANCELLED
            - SKIPPED
          type: string
        task_key:
          type: string
        task_type:
          type: string
      required:
        - id
        - task_key
        - task_type
        - attempt
        - state
        - reason
        - executor_type
        - pool
        - error
      type: object
...
    TriggerRequest:
      additionalProperties: true
      properties:
        inputs:
          additionalProperties: {}
          type: object
        labels:
          additionalProperties:
            maxLength: 256
            type: string
          maxProperties: 20
          type: object
      type: object
...
info:
  title: Sluice API
  version: "1"
openapi: 3.1.0
paths:
  /api/runner/v1/task-runs/{taskRunId}/artifacts/{name}:
    put:
      operationId: runnerPutArtifact
      parameters:
        - in: path
          name: taskRunId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: name
          required: true
          schema:
            maxLength: 200
            type: string
      requestBody:
        content:
          application/octet-stream:
            schema:
              contentMediaType: application/octet-stream
              format: binary
              type: string
        required: true
      responses:
        "204":
          description: Stored.
  /api/runner/v1/task-runs/{taskRunId}/bundle:
    get:
      operationId: runnerGetBundle
      parameters:
        - in: path
          name: taskRunId
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "200":
          content:
            application/gzip: {}
          description: Bundle.
  /api/runner/v1/task-runs/{taskRunId}/complete:
    post:
      operationId: runnerComplete
      parameters:
        - in: path
          name: taskRunId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunnerComplete"
        required: true
      responses:
        "204":
          description: No Content
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
  /api/runner/v1/task-runs/{taskRunId}/events:
    post:
      operationId: runnerPostEvents
      parameters:
        - in: path
          name: taskRunId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunnerEventBatch"
        required: true
      responses:
        "204":
          description: No Content
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
  /api/runner/v1/task-runs/{taskRunId}/heartbeat:
    post:
      operationId: runnerHeartbeat
      parameters:
        - in: path
          name: taskRunId
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerHeartbeatResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
  /api/runner/v1/task-runs/{taskRunId}/logs:
    post:
      operationId: runnerPostLogs
      parameters:
        - in: path
          name: taskRunId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunnerLogBatch"
        required: true
      responses:
        "204":
          description: No Content
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
  /api/runner/v1/task-runs/{taskRunId}/spec:
    get:
      operationId: runnerGetSpec
      parameters:
        - in: path
          name: taskRunId
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerSpec"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
...
  /api/v1/executions:
    get:
      operationId: listExecutions
      parameters:
        - description: Comma-separated execution states.
          explode: false
          in: query
          name: state
          schema:
            description: Comma-separated execution states.
            type: string
        - description: Namespace and its children.
          explode: false
          in: query
          name: namespace
          schema:
            description: Namespace and its children.
            type: string
        - description: Flow as <namespace>/<flow_id>.
          explode: false
          in: query
          name: flow
          schema:
            description: Flow as <namespace>/<flow_id>.
            type: string
        - explode: false
          in: query
          name: trigger_type
          schema:
            enum:
              - manual
              - schedule
              - webhook
              - flow
              - file
              - subflow
              - rerun
              - restart
            type: string
        - description: Label filter key=value. Repeat for several labels.
          explode: true
          in: query
          name: label
          schema:
            description: Label filter key=value. Repeat for several labels.
            items:
              type: string
            type: array
        - explode: false
          in: query
          name: from
          schema:
            format: date-time
            type: string
        - explode: false
          in: query
          name: to
          schema:
            format: date-time
            type: string
        - explode: false
          in: query
          name: sort
          schema:
            default: created
            enum:
              - created
              - duration
            type: string
        - description: Opaque cursor from next_cursor of the previous page.
          explode: false
          in: query
          name: cursor
          schema:
            description: Opaque cursor from next_cursor of the previous page.
            type: string
        - explode: false
          in: query
          name: limit
          schema:
            default: 50
            format: int64
            maximum: 200
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionList"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
  /api/v1/executions/{executionId}:
    get:
      operationId: getExecution
      parameters:
        - in: path
          name: executionId
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionDetail"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
...
  /api/v1/executions/{executionId}/events:
    get:
      operationId: streamExecutionEvents
      parameters:
        - in: path
          name: executionId
          required: true
          schema:
            format: uuid
            type: string
        - in: header
          name: Last-Event-ID
          schema:
            type: string
      responses:
        "200":
          content:
            text/event-stream: {}
          description: Event stream.
  /api/v1/executions/{executionId}/logs:
    get:
      operationId: getExecutionLogs
      parameters:
        - in: path
          name: executionId
          required: true
          schema:
            type: string
        - description: Task ID filter.
          explode: false
          in: query
          name: task
          schema:
            description: Task ID filter.
            type: string
        - description: Case-insensitive text filter.
          explode: false
          in: query
          name: search
          schema:
            description: Case-insensitive text filter.
            type: string
        - description: Opaque cursor from next_cursor of the previous page.
          explode: false
          in: query
          name: cursor
          schema:
            description: Opaque cursor from next_cursor of the previous page.
            type: string
        - explode: false
          in: query
          name: limit
          schema:
            default: 1000
            format: int64
            maximum: 5000
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogPage"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
...
  /api/v1/flows/{namespace}/{flowId}/executions:
    post:
      operationId: triggerFlow
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            maxLength: 128
            pattern: ^[a-z0-9]([a-z0-9-]*[a-z0-9])?(\.[a-z0-9]([a-z0-9-]*[a-z0-9])?)*$
            type: string
        - in: path
          name: flowId
          required: true
          schema:
            maxLength: 63
            pattern: ^[a-z0-9][a-z0-9-]*$
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionDetail"
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
...
  /api/v1/namespaces/{namespace}/run:
    post:
      operationId: runFile
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            maxLength: 128
            pattern: ^[a-z0-9]([a-z0-9-]*[a-z0-9])?(\.[a-z0-9]([a-z0-9-]*[a-z0-9])?)*$
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunFileRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionDetail"
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
...